requirements.txt
spotipy==2.23.0
mcp==1.0.0
pydantic==2.5.0
python-dotenv==1.0.0
